Removed the separated named exec. policy
The separated class for the named execution policy has been removed,
as the same functionality can be achieved using the same sycl
execution policy.
Functors can be used as names without having to specify a template
parameter for the policy - the name of the kernel will be the
name of the functor.