Update FindTriSYCL.cmake to latest version to remove a lot of warnings
Use -Wno-ignored-attributes to avoid a lot of warning with Boost.Compute and g++ version >= 6
The amount of warning was so high that it caused Travis CI to fail.

TODO: this FindTriSYCL.cmake should not be part of this repository but
should be used directly from the locally available triSYCL instead.